Renamed variant indexer package
