+++ agr_variant_indexer/src/main/java/org/alliancegenome/indexer/variant/Main.java
@@ -1,4 +1,4 @@
-package org.alliancegenome.variant_indexer;
+package org.alliancegenome.indexer.variant;
 
 import org.alliancegenome.core.config.ConfigHelper;
 import org.alliancegenome.core.filedownload.model.DownloadFileSet;
@@ -7,7 +7,7 @@ import org.alliancegenome.core.variant.config.VariantConfigHelper;
 import org.alliancegenome.es.index.site.schema.VariantMapping;
 import org.alliancegenome.es.index.site.schema.settings.VariantIndexSettings;
 import org.alliancegenome.es.util.IndexManager;
-import org.alliancegenome.variant_indexer.es.managers.*;
+import org.alliancegenome.indexer.variant.es.managers.*;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
