+++ agr_variant_indexer/src/main/java/org/alliancegenome/indexer/variant/es/managers/SourceDocumentCreationManager.java
@@ -1,4 +1,4 @@
-package org.alliancegenome.variant_indexer.es.managers;
+package org.alliancegenome.indexer.variant.es.managers;
 
 import java.util.concurrent.*;
